personal_data_pull/whatsapp_export/requirements.txt
numpy==1.22.0
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
textparser==0.23.0
